+++ presentation/src/main/java/org/gdsc/presentation/view/MainActivity.kt
@@ -1,26 +1,16 @@
 package org.gdsc.presentation.view
 
-import android.Manifest
 import android.animation.Animator
 import android.annotation.SuppressLint
-import android.content.Intent
-import android.content.pm.PackageManager
 import android.content.res.ColorStateList
-import android.net.Uri
 import android.os.Bundle
-import android.provider.Settings
 import android.view.MenuItem
 import android.view.View
-import android.view.animation.Animation
-import android.widget.Toast
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
-import androidx.core.content.ContextCompat
 import androidx.core.view.WindowCompat
 import androidx.core.view.isVisible
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import dagger.hilt.android.AndroidEntryPoint
 import org.gdsc.presentation.BaseActivity
 import org.gdsc.presentation.R
@@ -42,56 +32,7 @@ class MainActivity : BaseActivity() {
 
     private val myPageViewModel: MyPageViewModel by viewModels()
 
-    private val locationPermissionRequest = registerForActivityResult(
-        ActivityResultContracts.RequestMultiplePermissions()
-    ) { permissions ->
 
-        if (permissions.all { it.value }) {
-            Toast.makeText(this, "위치 권한이 허용되었습니다!", Toast.LENGTH_SHORT).show()
-        } else {
-            showLocationPermissionDialog()
-        }
-
-    }
-
-    private fun checkLocationRequest() {
-        if (ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.ACCESS_FINE_LOCATION
-            ) != PackageManager.PERMISSION_GRANTED || ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.ACCESS_FINE_LOCATION
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            locationPermissionRequest.launch(
-                arrayOf(
-                    Manifest.permission.ACCESS_FINE_LOCATION,
-                    Manifest.permission.ACCESS_COARSE_LOCATION
-                )
-            )
-        }
-    }
-
-    // TODO: change to JMT Dialog
-    private fun showLocationPermissionDialog() {
-        MaterialAlertDialogBuilder(this)
-            .setTitle("위치 권한 요청")
-            .setMessage("위치 권한 설정 창으로 이동할까요?")
-            .setPositiveButton("네") { _, _ ->
-
-                startActivity(
-                    Intent(
-                        Settings.ACTION_APPLICATION_DETAILS_SETTINGS,
-                        Uri.fromParts("package", packageName, null)
-                    )
-                )
-            }
-            .setNegativeButton("아니요") { _, _ ->
-                Toast.makeText(this, "위치 권한이 꼭 필요합니다.", Toast.LENGTH_SHORT).show()
-                finish()
-            }
-            .show()
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
 
@@ -99,7 +40,6 @@ class MainActivity : BaseActivity() {
 
         binding = ActivityMainBinding.inflate(layoutInflater)
 
-        checkLocationRequest()
         setContentView(binding.root)
         initBottomNavigationView()
         setToolbar()
